Updated input format documentation. Moved Input Formats to its own Concern trait to keep code base clean.
